Make Buildship work on Java 9
This change adds initial Java 9 compatibility to Buildship by
implementing the following changes:

 - Update Tooling API to 4.3-rc-3
 - Update wrapper to 4.3-rc-4
 - Update Gson to 2.7
 - Update Groovy to 2.4.12
 - Launch Eclipse with --add-modules=ALL-SYSTEM option on Java 9
 - Fix broken tests
 - Ignore tests that are using older Gradle versions
 - Use e47 build that support Java 9
 - GradleClasspathProvider filters unresolved rt classpath too

At the moment older Gradle versions cannot be used when Buildship
is executed on Java 9. With the new JDK versioning scheme we have
lost backward compatibility in this platform. We will try to make
it work but there's no guarantee that it will be done.